backend/requirements.txt
pymupdf
fastapi
requests
typer
python-dotenv